Still playing with the rules for proper sound levels
I'm still trying to get the recording function working correctly.
The logic seems to be that you start recording. If the audio level
has not gone above the threshold after n seconds, the threshold
may be too high. Set the threshold to the average audio level plus
five.
If the audio level has not gone below the threshold after n seconds,
the audio level may be too low and Jasper may continue to listen
forever, unable to interrupt. Again, set the threshold to the
average audio level plus five.