Change interface from a list of responses to one response at a time
Changed the interface from listing responses, and often taking several
seconds to render, to just presenting one response at a time. Also
added a new Result "Unclear" in addition to "Correct", "Modify" and
"Nothing" for samples that contain speech, but it is muffled or garbled
or otherwise should be ignored.